pyproject.toml
[tool.poetry]
name = "mocap"
version = "0.1.0"
description = ""
authors = ["michal_niedbala <mniedbala99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.1"
opencv-python = "^4.8.0.74"
pypylon = "^2.2.0"
paho-mqtt = "^1.6.1"
pyyaml = "^6.0"
matplotlib = "^3.7.2"
pyzmq = "^25.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
